requirements.txt
python-dotenv==1.0.0
pytz==2023.3
tweepy==4.14.0